Add OG image support: superadmin upload, owner picker, link preview thumbnail
- Schema: add og_title/og_description/og_image to sessions; add og_images table
- /api/og-images GET: public list of available images for owners
- /api/sa/og-images GET/POST: SA-only image upload to Supabase Storage + DB record
- /api/sa/og-images/[id] DELETE: remove image from storage and DB
- /app/sa/(panel)/og-images: SA panel to upload/manage OG images (grid view)
- SaSidebar: add "OG Images" nav item
- Dashboard settings modal: add og_title, og_description, og_image picker fields
- /app/p/[slug]/page.tsx: split into server wrapper + CustomerPageClient.tsx;
  generateMetadata now reads og_title/og_description/og_image and emits
  og:image and twitter:card=summary_large_image when image is set
- Sessions PUT API: accepts og_image field
