+++ app/api/og-images/route.ts
@@ -1,0 +1,12 @@
+import { NextResponse } from "next/server";
+import { supabase } from "@/lib/supabase";
+
+export async function GET() {
+  const { data, error } = await supabase
+    .from("og_images")
+    .select("id, url, name, created_at")
+    .order("created_at", { ascending: false });
+
+  if (error) return NextResponse.json({ error: error.message }, { status: 500 });
+  return NextResponse.json(data || []);
+}
